UI component added, remove comments
