+++ Helper/Helper/HelpSeekerViews/Profile.swift
@@ -60,20 +60,3 @@ struct Profile_Previews: PreviewProvider {
         Profile(helpseekerName: .constant(""))
     }
 }
-
-struct SettingsView: View {
-    var name: String
-    var body: some View {
-        Button(action: {
-            
-        }){
-            HStack {
-                Text(name)
-                Spacer(minLength: 15)
-            }.padding()
-                .font(.headline)
-                .foregroundColor(Color.black.opacity(0.6))
-        }
-    }
-}
-
