+++ Helper/Helper/VolunteerViews/VolunteerDashboard.swift
@@ -23,7 +23,7 @@ struct VolunteerDashboard: View {
     @State var assistance: [Task] = []
     @State var transport: [Task] = []
     @State var others: [Task] = []
-    // @State var taskStatus: Int16
+
     // get task of current user
     func getTaskInfo() {
         DispatchQueue.main.async {
@@ -157,141 +157,3 @@ struct VolunteerDashboard_Previews: PreviewProvider {
     }
 }
 
-struct ImageSlideShow: View{
-    private let timer = Timer.publish(every: 5, on: .main, in: .common).autoconnect()
-    @State private var currentIndex = 0
-    
-    var body: some View{
-        GeometryReader{ geometry in
-            TabView(selection: $currentIndex){
-                ForEach(0..<4){image in
-                    Image("\(image)")
-                        .resizable()
-                        .scaledToFill()
-                        .overlay(Color.black.opacity(0.06))
-                        .tag(image)
-                }
-            }.tabViewStyle(PageTabViewStyle())
-                .clipShape(RoundedRectangle(cornerRadius: 10))
-                .onReceive(timer, perform: {_ in
-                    withAnimation{
-                        currentIndex = currentIndex < 4 ? currentIndex + 1 :0
-                        
-                    }
-                })
-        }
-    }
-}
-
-struct CategoriesView: View {
-    
-    var categoryName: String
-    var numberOfTasks: String
-    var ImageName: String
-    var body: some View {
-        
-        ZStack{
-            RoundedRectangle(cornerRadius: 10)
-                .fill(.white)
-                .frame(width: 110, height: 140)
-                .shadow(radius: 5)
-            VStack(alignment: .leading){
-                Text(categoryName)
-                    .font(.headline)
-                    .fontWeight(.medium)
-                    .foregroundColor(.primary)
-                    .padding(.top,25)
-                Text(numberOfTasks)
-                    .font(.subheadline)
-                    .foregroundColor(.primary)
-                Image(ImageName)
-                    .resizable()
-                    .frame(width: 100, height: 75,alignment: .center)
-                    .padding(.bottom,20)
-            }
-        }
-    }
-}
-
-struct OngoingTaskCard: View {
-    @FetchRequest(entity: Task.entity(), sortDescriptors: [NSSortDescriptor(keyPath: \Task.title, ascending: true)]) var taskResults: FetchedResults<Task>
-    @Environment(\.managedObjectContext) var context
-    @State var refresh: Bool = false
-    @State var taskTitle: String
-    @State var helpseeker: String
-    @State var location: String
-    @State var time: Date?
-    @State var date: Date?
-    @State var desc: String
-    @State var need: String
-    @State var chronic: String
-    @State var allergies: String
-    @State var id: Int16
-    
-    
-    @State var taskId: Int16 = 0
-    var userInfo: [User]  = []
-    var body: some View {
-        NavigationLink(destination: TaskDetailView(taskTitle: taskTitle, helpseeker: helpseeker, location: location,
-                                                   time: time, desc: desc, need: need,
-                                                   chronic: chronic, allergies: allergies, id: id)){
-            
-            ZStack{
-                RoundedRectangle(cornerRadius: 10)
-                    .fill(.white)
-                    .shadow(radius: 5)
-                    .frame(width: 340, height: 150)
-                
-                VStack(alignment:.leading,spacing:10){
-                    HStack{
-                        Text(taskTitle)
-                            .font(.headline)
-                            .fontWeight(.semibold)
-                            .foregroundColor(.black)
-                        Spacer()
-                        Text(time!.formatted(date: .numeric, time: .omitted))
-                            .font(.subheadline)
-                            .fontWeight(.semibold)
-                            .foregroundColor(Color("Primary"))
-                    }.padding(.horizontal)
-                    Label("\(helpseeker)", systemImage: "person")
-                        .padding(.horizontal)
-                        .foregroundColor(.black)
-                    Label("\(location)", systemImage: "mappin")
-                        .padding(.horizontal)
-                        .foregroundColor(.black)
-                    HStack{
-                        Label(time!.formatted(date: .omitted, time: .complete), systemImage: "clock")
-                            .font(.system(size: 14))
-                            .foregroundColor(.secondary)
-                        Spacer()
-                        
-                        Button(action: {
-                            
-                            self.taskId = id
-                            if(taskId != 0){
-                                if let index = taskResults.firstIndex(where: {$0.id == taskId}) {
-                                    taskResults[index].status = 1
-                                }
-                                do {
-                                    try context.save()
-                                } catch {
-                                    print(error)
-                                }
-                            }
-                        })
-                        {
-                            Text("Mark As Done")
-                                .font(.subheadline)
-                                .frame(width: 110, height: 30)
-                                .background(Color("Primary"))
-                                .foregroundColor(.white)
-                                .cornerRadius(6)
-                        }
-                    }
-                    .padding(.bottom,5)
-                }
-            }
-        }
-    }
-}
